Fix in question & test management controllers, minor fixes on pages

+++ project/OnlineExamSystem/src/java/controller/oesadmin/question/RemoveController.java
@@ -11,6 +11,7 @@ import javax.servlet.annotation.ServletSecurity;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import model.Account;
 import model.Question;
 import util.servlet.ManagedServlet;
 
@@ -25,10 +26,19 @@ public class RemoveController extends ManagedServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
+        Account currentUser = (Account) request.getSession().getAttribute("currentUser");
         String questionID = request.getParameter("id");
         QuestionManager qm = new QuestionManager();
         Question question = qm.getQuestion(Long.parseLong(questionID));
-        boolean del = qm.deleteQuestion(question);
+        if (question == null) {
+            response.sendError(400, "Invalid question ID.");
+            return;
+        }
+        if (!question.getOwner().equals(currentUser)) {
+            response.sendError(403, "You do not have permission to delete the question.");
+            return;
+        }
+        qm.deleteQuestion(question);
         redirect(response, getServletURL(QuestionController.class));
     }
 }
